rename syncer package to syncers
